split most of the models and storage updating logic into pool_models.rs
