+++ shared/src/balancer.rs
@@ -1,3 +1,4 @@
 pub mod event_handler;
 mod info_fetching;
 pub mod pool_fetching;
+mod pool_models;
