+++ shared/src/balancer/pool_fetching.rs
@@ -1,51 +1,18 @@
 use anyhow::Result;
 use model::TokenPair;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 
-use crate::balancer::event_handler::{BalancerPoolRegistry, RegisteredWeightedPool};
+use crate::balancer::{
+    event_handler::BalancerPoolRegistry,
+    pool_models::{RegisteredWeightedPool, WeightedPool},
+};
 use crate::pool_fetching::{handle_contract_error, MAX_BATCH_SIZE};
 use crate::recent_block_cache::Block;
 use crate::Web3;
 use contracts::BalancerV2Vault;
 use ethcontract::batch::CallBatch;
 use ethcontract::errors::MethodError;
-use ethcontract::{BlockId, Bytes, H160, H256, U256};
-
-pub struct PoolTokenState {
-    pub balance: U256,
-    pub weight: U256,
-    pub scaling_exponent: u8,
-}
-
-pub struct WeightedPool {
-    pub pool_id: H256,
-    pub pool_address: H160,
-    pub reserves: HashMap<H160, PoolTokenState>,
-}
-
-impl WeightedPool {
-    fn new(pool_data: RegisteredWeightedPool, balances: Vec<U256>) -> Self {
-        let mut reserves = HashMap::new();
-        // We expect the weight and token indices are aligned with balances returned from EVM query.
-        // If necessary we would also pass the tokens along with the query result,
-        // use them and fetch the weights from the registry by token address.
-        for (i, balance) in balances.into_iter().enumerate() {
-            reserves.insert(
-                pool_data.tokens[i],
-                PoolTokenState {
-                    balance,
-                    weight: pool_data.normalized_weights[i],
-                    scaling_exponent: pool_data.scaling_exponents[i],
-                },
-            );
-        }
-        WeightedPool {
-            pool_id: pool_data.pool_id,
-            pool_address: pool_data.pool_address,
-            reserves,
-        }
-    }
-}
+use ethcontract::{BlockId, Bytes, H160, U256};
 
 #[async_trait::async_trait]
 pub trait WeightedPoolFetching: Send + Sync {
